+++ CHANGELOG.md
@@ -4,3 +4,7 @@ All notable changes to the "papyrus-tools" extension will be documented in this 
 
 ## [0.1.0] - 2025-10-08
 - Initial release with syntax, snippets, completions, and hovers
+- Added webpack bundling
+- Added document symbols, go-to definition
+- Added basic diagnostics (If/EndIf, While/EndWhile)
+- Added compile command and settings